Make LCD reconfigure V3 users when auth data changes.
`CommandGeneratorLcdConfigurator` configures auth data and caches based
on `(username, security engine id)`. If the authentication data is
changed, such as via `usmUserAuthKeyChange`, the cache is not updated
because neither the username nor engine ID changed. Despite new data
being passed in, this results in the old data being used and a "Wrong
SNMP PDU digest" error indication being returned.

Change the LCD to detect a changed authKey, authProtocol, privKey, or
privProtocol and reconfigure the V3 user as appropriate.